+++ src/application-base/private/getBrowserLocale.js
@@ -24,6 +24,13 @@ const filterLanguages = () => {
   if (navigator.languages && navigator.languages.length > 0) {
     for (let index = 0; index < navigator.languages.length; index += 1) {
       const locale = navigator.languages[index];
+      const fallbackLocale = locale.split('-').length > 1 ? locale.split('-')[0] : false;
+
+      if(fallbackLocale){
+        if (isSupported(fallbackLocale)) {
+          return fallbackLocale;
+        }
+      }
 
       if (isSupported(locale)) {
         return locale;